requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.2.8
django-ckeditor==6.7.3
django-js-asset==3.1.2
psycopg2-binary==2.9.11
sqlparse==0.5.3
tzdata==2025.2
